refactor: consolidate SSR logger + http-injection core in shared tree
Per review on #485. Moves two more pieces out of the node runtime and
into runtimes/javascript so bun can consume them verbatim:

- src/ssr/logger.mjs — node had a cls-hooked copy, bun had an
  AsyncLocalStorage .ts port. Converge on AsyncLocalStorage (stable
  under both runtimes), drop cls-hooked from node's package.json, and
  move the single file into the shared tree.

- src/ssr/http-injection.mjs — the overrideEmit/wrapServer/
  wrapCreateServer/wrapHttp helpers. These were byte-identical between
  node's injections.mjs and bun's injections.ts (only the hook/preload
  registration differed). Shared file exports the helpers; node's
  injections.mjs shrinks to import-in-the-middle + require-in-the-
  middle registration only. Bun will consume the same helpers in #485.

No behaviour changes — the per-request contract (health/timings
endpoints, secret auth, log-id header, safe-timeout, error capture)
is identical. Logger callers that used to do
loggingNamespace.run(fn) + .set("id", id) now use
loggingNamespace.run({ id }, fn) (AsyncLocalStorage shape).
