+++ runtimes/node/versions/latest/package.json
@@ -12,7 +12,6 @@
   "license": "ISC",
   "dependencies": {
     "@vercel/nft": "^0.29.2",
-    "cls-hooked": "^4.2.2",
     "express": "^4.21.2",
     "import-in-the-middle": "^1.14.4",
     "micro": "^9.3.4",
